requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2016.2.28
cffi==1.14.1
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.1
flask_googlemaps==0.4.0
flask-googlemaps==0.4.0